Initial sketch of VisibleState API
The idea here is that we expose just one API for communicating
the visible state shown in the user interface.

The Hearts engine will construct VisibleState and apply it to
the CardsWebServer whenever the visible state changes.

VisibleState is intentionally immutable. But I imagine I will
add methods to apply a state change, which will create a new
immutable VisibleState from the previous state and an operation.